Remove Joda Time and use java.time package instead
Joda Time has been superseded by the Java 8 java.time package
which has more or less the same API. This commit removes all
usage of Joda Time and replaces it with the java.time classes.
This is done so that we can remove the dependency on Joda Time.

+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/compose/state/ConnectUiState.kt
@@ -11,7 +11,7 @@ data class ConnectUiState(
     val showLocation: Boolean,
     val inAppNotification: InAppNotification?,
     val deviceName: String?,
-    val daysLeftUntilExpiry: Int?,
+    val daysLeftUntilExpiry: Long?,
     val isPlayBuild: Boolean,
 ) {
